pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rt97l-tool"
version = "0.1.0"
description = "Linux programmer for Retevis RT97L / CX-525 GMRS repeaters"
requires-python = ">=3.9"
dependencies = [
    "pyserial>=3.5",
    "textual>=0.80",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
rt97l = "rt97l.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
